Improve DIM2icicle_updateHitResponse to 99.04%
Force the materialized ~K (li -65; and) form for the SEQUENCE_FLAG_0040 clear
via ~(u64)MACRO; the plain u32 mask emitted rlwinm. Residual is register-
permutation branch-displacement noise.

+++ src/main/dll/DIM/dim2icicle.c
@@ -553,7 +553,7 @@ void DIM2icicle_updateHitResponse(GameObject *obj, DIMbossRuntime *runtime)
   }
   hitResult = ObjHits_GetPriorityHit(objIndex, &hitId, &hitType, &hitVolume);
   if (hitResult != 0) {
-    gDIMbossSequenceFlags = gDIMbossSequenceFlags & ~DIMBOSS_SEQUENCE_FLAG_0040;
+    gDIMbossSequenceFlags = gDIMbossSequenceFlags & ~(u64)DIMBOSS_SEQUENCE_FLAG_0040;
     if (objectRuntime->phase == DIMBOSS_PHASE_LAUNCH_LIFT) {
       if ((gDIMbossSequenceFlags & DIMBOSS_SEQUENCE_FLAG_TONSIL_GUARD_ACTIVE) == 0 ||
           hitType != 2) {
